% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdfsFunctions.R
\name{copyMergeToLocal}
\alias{copyMergeToLocal}
\title{read data from hdfs as a dataframe in R}
\usage{
copyMergeToLocal(hdfsUri, src, ...)
}
\arguments{
\item{hdfsUri}{the namenode of your hdfs, just as hdfs://192.163.8.1:9000}

\item{src}{the file or directory in your hdfs. if it is a directory, the funciton will get all files in the directroy, an read them as a dataframe, and if there are other directories in src, they will not read.}

\item{...}{the arguments can be pass to read.table.}

\item{sep}{the separator character.}
}
\description{
read data from hdfs as a dataframe in R
}

